+++ Tests/Integration/DSCResources/McAfee.config.ps1
@@ -26,6 +26,10 @@ configuration McAfee_config
         $SkipRuleType,
 
         [Parameter()]
+        [string[]]
+        $SkipRuleCategory,
+
+        [Parameter()]
         [hashtable]
         $Exception,