+++ Tests/Integration/DSCResources/Vsphere.config.ps1
@@ -34,6 +34,10 @@ configuration Vsphere_config
         $SkipRuleType,
 
         [Parameter()]
+        [string[]]
+        $SkipRuleCategory,
+
+        [Parameter()]
         [object]
         $OrgSettings,